move BlurPoolHparams
